Set up merge files with webpack
