--- index.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="index.css">
-    <title>Meals</title>
-    <!--<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">-->
-</head>
-<body>
-    <header>
-        <p class="title">Available Meals </p>
-        <ul class="nav">
-            <li class="tab"><a href="#">Home</a></li>
-            <li><hr></li>
-            <li class="tab"><a href="#">About</a></li>
-            <li><hr></li>
-            <li class="tab"><a href="#">Contact</a></li>
-        </ul>
-    </header>
-	<ul id="meals-list"></ul>
-	<div id="meal-details"></div>
-	<script src="app.js"></script>
-</body>
-</html>
