+++ src/index.html
@@ -4,13 +4,22 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Wbpack Exercise</title>
+    <script type="module" src="./index.js" defer></script>
 </head>
 <body>
     <header>
-
+        <p class="title">Available Meals </p>
+        <ul class="nav">
+            <li class="tab"><a href="#">Home</a></li>
+            <li><hr></li>
+            <li class="tab"><a href="#">About</a></li>
+            <li><hr></li>
+            <li class="tab"><a href="#">Contact</a></li>
+        </ul>
     </header>
     <main>
-
+        <ul id="meals-list"></ul>
+        <div id="meal-details"></div>
     </main>
     <footer>
         <div class="footer-text">Created by Matee Safi and Therese Tuyisabe under CC licence</div>
