+++ src/style.css
@@ -1,7 +1,3 @@
-.hello {
-  color: red;
-}
-
 .footer-text {
   display: block;
   padding: 3em;
@@ -11,4 +7,86 @@
   height: 2em;
   width: 100%;
   background-color: gray;
+}
+
+header {
+  width: 100%;
+  display: flex;
+  height: 4vw;
+  background-color: rgba(228, 18, 18, 0.915);
+  padding: 1vh 10%;
+  gap: 41vw;
+  align-items: center;
+}
+
+.title {
+  color: rgb(15, 14, 14);
+  font-size: 2.5vw;
+  font-weight: 700;
+}
+
+.nav {
+  list-style-type: none;
+  color: rgb(27, 2, 2);
+  display: flex;
+  gap: 1.2vw;
+  align-items: center;
+}
+
+body {
+  width: 100vw;
+
+  /* background-image: url('./src/images/'); */
+  background-size: contain;
+  background-attachment: fixed;
+  overflow-x: hidden;
+}
+
+/* Style for the meals list */
+#meals-list {
+  width: 100%;
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-between;
+  list-style: none;
+  padding: 10px;
+  gap: 1vw;
+}
+
+/* Style for each meal item */
+#meals-list li {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: start;
+  width: calc(33.33% - 20px);
+  margin-bottom: 20px;
+  position: relative;
+}
+
+#meals-list li img {
+  display: block;
+  margin-bottom: 10px;
+  width: 90%;
+  height: 100%;
+}
+
+#meals-list li .meal-info {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  width: 100%;
+}
+
+#meals-list li .meal-info span {
+  font-weight: bold;
+}
+
+#meals-list li .like-btn {
+  background-color: #fff;
+  border: none;
+  cursor: pointer;
+  outline: none;
+  position: absolute;
+  right: 10px;
 }
